Scheduled announcements
Added some scheduled announcements to tell players when the weekend bonus starts and ends.

I couldn't find a reliable way of doing this using the TaskScheduler, so I just made this for now. If I figure out a better way of doing it, I'll change it. If I find out this is plain horrible, I'll remove it.